Use hooks instead of middleware for auth
Use an onRequest hook to perform authentication logic. This ensures
that even when sending a 401, the CORs headers are still set.

closes #294
